+++ src/javascripts/components/singleBoard/singleBoard.js
@@ -2,10 +2,11 @@ import $ from 'jquery';
 // import firebase from 'firebase/auth';
 
 import pinsData from '../../helpers/data/pinsData';
+import aPin from '../pins/pins';
 // import boardData from '../../helpers/data/boardsData';
 
 import './singleBoard.scss';
-import utilities from '../../helpers/utilities';
+// import utilities from '../../helpers/utilities';
 
 const backToBoards = (e) => {
   e.preventDefault();
@@ -15,26 +16,11 @@ const backToBoards = (e) => {
 
 const showOneBoard = (boardId) => {
   pinsData.getMyPins(boardId)
-    .then((pins) => {
-      let domString = '<div class="row justify-content-between"><h1>Pins</h1>';
-      domString += '<button class="btn btn-success" id="all-boards">Back to Boards</button></div>';
-      domString += '<div class="row">';
-      pins.forEach((pin) => {
-        domString += `
-        <div class="card col-4">
-          <img src="${pin.imageUrl}" class="card-img-top" alt="${pin.pinName}">
-          <div class="card-body">
-            <h5 class="card-title">${pin.pinName}</h5>
-            <button class="btn btn-danger delete" id="${pin.pinId}-del-${boardId}">Delete</button>
-          </div>
-        </div>
-      `;
-      });
-      domString += '</div>';
-      utilities.printToDom('single-board', domString);
+    .then(() => {
       $('#single-board').on('click', '#all-boards', backToBoards);
       $('#boards').addClass('d-none');
       $('#single-board').removeClass('d-none');
+      $('body').on('click', '.delete', aPin.deletePin);
     })
     .catch((error) => console.error(error));
 };
